moveonly: add mp/type-interface.h
