+++ example/types.h
@@ -6,6 +6,7 @@
 #define EXAMPLE_TYPES_H
 
 #include <mp/type-context.h>
+#include <mp/type-interface.h>
 #include <mp/type-string.h>
 
 #endif // EXAMPLE_TYPES_H
